refactor(grpc_client): move other main types into util
